npm init, packages, index, setup
