+++ package.json
@@ -1,19 +1,31 @@
 {
-  "name": "chris-reilly-project",
+  "name": "node-blog",
   "version": "1.0.0",
-  "description": "Chris Reilly Project",
+  "description": "Lambda backend w01d03",
   "main": "index.js",
-  "repository": "https://github.com/duckets615/Node-Blog.git",
-  "author": "duckets615 <33557104+duckets615@users.noreply.github.com>",
-  "license": "MIT",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "server": "nodemon index.js",
+    "supervisor": "supervisor index.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/duckets615/Node-Blog.git"
+  },
+  "author": "Chris Reilly",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/duckets615/Node-Blog/issues"
+  },
+  "homepage": "https://github.com/duckets615/Node-Blog#readme",
   "dependencies": {
-    "cors": "^2.8.4",
-    "d": "^1.0.0",
-    "express": "^4.16.3",
-    "helmet": "^3.13.0",
-    "knex": "^0.15.2",
+    "cors": "^2.8.5",
+    "express": "^4.16.4",
+    "helmet": "^3.16.0",
+    "knex": "^0.16.3",
     "morgan": "^1.9.1",
-    "nodeman": "^1.1.2",
-    "sqlite3": "^4.0.2"
+    "nodemon": "^1.18.10",
+    "sqlite3": "^4.0.6",
+    "supervisor": "^0.12.0"
   }
 }
